Refactor ProteinAnalyzer to improve modularity and add phi/psi dihedral angle and secondary structure calculation functionality
